requirements.txt
numpy==1.18.1
scikit-image==0.16.2
scikit-learn==0.22.1
matplotlib==3.1.3
numba==0.49.1
tqdm==4.46.0
attrs==19.3.0
streamlit==0.62.1
imageio==2.8.0
pandas==1.0.5
altair==4.1.0
scipy==1.4.1